Move color-based util fns into separate .py file
