--- great_tables/_data_color/__init__.py
@@ -1,3 +1,0 @@
-from .base import data_color
-
-__all__ = ["data_color"]
