+++ tests/test_formats.py
@@ -7,7 +7,7 @@ import re
 from great_tables import GT
 from great_tables.data import exibble
 from great_tables.gt import _get_column_of_values
-from great_tables._data_color.base import _html_color
+from great_tables._utils_color import _html_color
 from great_tables._utils_render_html import create_body_component_h
 from great_tables._formats import (
     _format_number_fixed_decimals,
